+++ cypress/e2e/goVerbs.cy.js
@@ -5,28 +5,15 @@ describe("go verbs", () => {
     cy.startGame();
   });
 
-  it("requires player to be holding the device before going west", () => {
-    cy.say("w", "have to be holding the strange device");
-    cy.say("take device", "the strange device");
-    cy.say("w", "Going west");
-    cy.choose("Next", "beautiful forest green", { global: true });
-  });
-
-  it("requires a door to be open before the player can go through it", () => {
+  it("allows the player to go through a door, possibly with an alias, if conditions are met", () => {
     cy.say("take device", "the strange device");
     cy.say("w", "Going west");
     cy.choose("Next", "beautiful forest green", { global: true });
     cy.say("go through door", "door is closed");
-  });
-
-  it("allows the player to go through a door if the condition is met", () => {
-    cy.say("take device", "the strange device");
-    cy.say("w", "Going west");
-    cy.choose("Next", "beautiful forest green", { global: true });
     cy.say("open door", "opens relatively easily");
     cy.say("drop device", "device on the floor");
     cy.say("go through door", "have to be holding the strange device");
     cy.say("take device");
-    cy.say("go through door", "step into the green wall");
+    cy.say("transit door", "step into the green wall");
   });
 });